Moving renderer back to the same directory it was in before
